+++ Tests/Common/VideoPlayerSpec.swift
@@ -108,10 +108,10 @@ class VideoPlayerSpec: QuickSpec {
                 when(mock).setOnSkipForwardButtonTapped(any()).thenDoNothing()
                 when(mock).setOnTimeSliderSlide(any()).thenDoNothing()
                 when(mock).setOnTimeSliderRelease(any()).thenDoNothing()
-                when(mock).setControlViewVisibility(visible: any(), animated: any()).then { tuple in
+                when(mock).setControlViewVisibility(visible: any(), withAnimationDuration: any()).then { tuple in
                     controlViewHasAlpha = tuple.0
                 }
-                when(mock).setInfoViewVisibility(visible: any(), animated: any()).then { tuple in
+                when(mock).setInfoViewVisibility(visible: any(), withAnimationDuration: any()).then { tuple in
                     infoViewHasAlpha = tuple.0
                 }
                 when(mock).setPlayButtonTo(state: any()).thenDoNothing()
@@ -423,65 +423,65 @@ class VideoPlayerSpec: QuickSpec {
 
         describe("info layer") {
             it("shows the info layer when there is no stream") {
-                verify(self.mockView, times(0)).setInfoViewVisibility(visible: true, animated: any())
+                verify(self.mockView, times(0)).setInfoViewVisibility(visible: true, withAnimationDuration: any())
 
                 self.videoPlayer.event = EntityBuilder.buildEvent(withStream: false, withStreamURL: false)
 
-                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, animated: any())
+                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, withAnimationDuration: any())
             }
 
             it("shows the info layer when there is no stream url") {
-                verify(self.mockView, times(0)).setInfoViewVisibility(visible: true, animated: any())
+                verify(self.mockView, times(0)).setInfoViewVisibility(visible: true, withAnimationDuration: any())
 
                 self.videoPlayer.event = EntityBuilder.buildEvent(withStream: true, withStreamURL: false)
 
-                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, animated: any())
+                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, withAnimationDuration: any())
             }
 
             it("does not allow info layer dismissal when there is no stream") {
                 self.videoPlayer.event = EntityBuilder.buildEvent(withStream: true, withStreamURL: false)
 
-                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, animated: any())
+                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, withAnimationDuration: any())
 
                 controlViewTapped?()
 
                 // Verify the count did not go up.
-                verify(self.mockView, times(0)).setInfoViewVisibility(visible: false, animated: any())
+                verify(self.mockView, times(0)).setInfoViewVisibility(visible: false, withAnimationDuration: any())
             }
 
             it("dismisses info layer on controlview tapped when there is a stream") {
                 self.videoPlayer.event = EntityBuilder.buildEvent(withStream: true, withStreamURL: true)
 
-                verify(self.mockView, times(1)).setInfoViewVisibility(visible: false, animated: any())
+                verify(self.mockView, times(1)).setInfoViewVisibility(visible: false, withAnimationDuration: any())
 
                 infoButtonTapped?()
 
-                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, animated: any())
+                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, withAnimationDuration: any())
 
                 controlViewTapped?()
 
-                verify(self.mockView, times(2)).setInfoViewVisibility(visible: false, animated: any())
+                verify(self.mockView, times(2)).setInfoViewVisibility(visible: false, withAnimationDuration: any())
             }
 
             it("hides the info layer when a stream url appears on a previously loaded event") {
                 self.videoPlayer.event = EntityBuilder.buildEvent(withRandomId: false, withStream: true, withStreamURL: false)
 
-                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, animated: any())
+                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, withAnimationDuration: any())
 
                 self.videoPlayer.event = EntityBuilder.buildEvent(withRandomId: false, withStream: true, withStreamURL: true)
 
-                verify(self.mockView, times(1)).setInfoViewVisibility(visible: false, animated: any())
+                verify(self.mockView, times(1)).setInfoViewVisibility(visible: false, withAnimationDuration: any())
             }
 
             it("does not show info layer again when the same event/stream is updated") {
                 let event = EntityBuilder.buildEvent(withRandomId: false, withStream: true, withStreamURL: false)
                 self.videoPlayer.event = event
 
-                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, animated: any())
+                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, withAnimationDuration: any())
 
                 self.videoPlayer.event = event
 
-                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, animated: any())
+                verify(self.mockView, times(1)).setInfoViewVisibility(visible: true, withAnimationDuration: any())
             }
         }
 
@@ -511,11 +511,13 @@ class VideoPlayerSpec: QuickSpec {
 
             it("calls delegate with slider update") {
                 class Delegate: VideoPlayerDelegate {
+
                     var called = false
 
                     func playerDidUpdatePlaying(player: VideoPlayer) {}
                     func playerDidUpdateState(player: VideoPlayer) {}
                     func playerDidUpdateFullscreen(player: VideoPlayer) {}
+                    func playerDidUpdateControlVisibility(toVisible: Bool, withAnimationDuration: Double, player: VideoPlayer) {}
                     func playerDidUpdateTime(player: VideoPlayer) {
                         called = true
                     }