+++ CHANGELOG.md
@@ -2,8 +2,17 @@
 All notable changes to this project will be documented in this file.
 `MLSSDK` adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.2.1](https://github.com/MyCujoo/mls-ios/releases/tag/1.2.1)
+Released on 2021-01-??.
+
+#### Added
+- A new method for the VideoPlayerDelegate that indicates that the visibility of the control layer was changed.
+
+#### Fixed
+- Hide the Airplay button while Chromecasting
+
 ## [1.2.0](https://github.com/MyCujoo/mls-ios/releases/tag/1.2.0)
-Released on 2020-01-14.
+Released on 2021-01-14.
 
 #### Added
 - Added alpha support for Chromecast